-Added a bunch of comments. Not quite done yet. -Removed some unused variables, renamed some others.

+++ src/lib/standard_summary_stats.c
@@ -1,3 +1,9 @@
+/**
+ * @file standard_summary_stats.c
+ * @brief User-defined functions to compute standard summary statistics.
+ * @author <Pierre-Luc.Drouin@drdc-rddc.gc.ca>, Defence Research and Development Canada Ottawa Research Centre.
+ */
+
 #include "standard_summary_stats.h"
 
 void std_stats_init(struct sim_vars* sv, struct std_summary_stats* stats)
@@ -6,7 +12,7 @@ void std_stats_init(struct sim_vars* sv, struct std_summary_stats* stats)
   for(i=sv->nlayers-1; i>=0; --i) {
     sv->iis[i].dataptr=malloc(sizeof(double));
   }
-  stats->npers=sv->pars.tmax+1;
+  stats->npers=(int)sv->pars.tmax+1;
   stats->inf_timeline=(uint32_t*)malloc(stats->npers*sizeof(uint32_t));
   stats->totinf_timeline=(uint32_t*)malloc(stats->npers*sizeof(uint32_t));
 }
